add bridge process compose
