+++ justfile
@@ -8,6 +8,8 @@ mcr-contract-tests:
     cd ./protocol-units/settlement/mcr/contracts && forge test
 mcr-client RUNTIME FEATURES *ARGS:
     ./scripts/movement/run mcr-client {{ RUNTIME }} {{ FEATURES }} {{ ARGS }}
+bridge RUNTIME FEATURES *ARGS:
+    ./scripts/movement/run bridge {{ RUNTIME }} {{ FEATURES }} {{ ARGS }}
 build-push-container IMAGE:
     ./scripts/movement/build-push-image {{ IMAGE }}
 container-tests:
