+++ process-compose/bridge/process-compose.bridge.yml
@@ -1,0 +1,20 @@
+version: "3"
+
+processes:
+    
+  bridge:
+    command: |
+      bridge-service
+    env:
+      RUST_LOG: info
+    readiness_probe:
+      initial_delay_seconds: 5
+      exec:
+        command: echo "true"
+    depends_on:
+      build:
+        condition: process_completed_successfully
+      bridge_setup:
+        condition: process_healthy
+      suzuka-full-node:
+        condition: process_healthy
