+++ process-compose/bridge/process-compose.build.yml
@@ -1,0 +1,11 @@
+version: "3"
+
+environment:
+
+processes:
+
+  build:
+    command: |
+      ./scripts/services/bridge/build
+    availability:
+      restart: exit_on_failure
