+++ process-compose/bridge/process-compose.celestia-local.yml
@@ -1,0 +1,71 @@
+version: "3"
+
+environment:
+
+processes:
+
+  setup:
+    command: |
+      export ETH_RPC_CONNECTION_PROTOCOL=http
+      export ETH_RPC_CONNECTION_HOSTNAME=0.0.0.0
+      export ETH_RPC_CONNECTION_PORT=8090
+      export MAYBE_RUN_LOCAL=true
+      export MAYBE_DEPLOY_MCR=true
+      suzuka-full-node-setup
+    depends_on:
+      build:
+        condition: process_completed_successfully
+    readiness_probe:
+      initial_delay_seconds: 3
+      exec:
+        command: echo "true"
+
+  celestia-node-appd:
+    command: |
+      m1-da-light-node-celestia-appd
+    readiness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: curl http://0.0.0.0:26657
+    liveness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: curl http://0.0.0.0:26657
+    depends_on:
+      setup:
+        condition: process_healthy
+
+  celestia-bridge:
+    command: |
+      m1-da-light-node-celestia-bridge
+    readiness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: curl http://0.0.0.0:26658
+    liveness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: curl http://0.0.0.0:26658
+    depends_on:
+       celestia-node-appd:
+        condition: process_healthy
+
+  celestia-light-node:
+    command: |
+      sleep 999999999
+      echo "started"
+    readiness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: |
+          echo "true"
+    liveness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: |
+          echo "true"
+    depends_on:
+      celestia-node-appd:
+        condition: process_healthy
+      celestia-bridge:
+        condition: process_healthy
