+++ process-compose/bridge/process-compose.setup.yml
@@ -1,0 +1,23 @@
+version: "3"
+
+processes:
+
+  setup:
+    command: |
+      suzuka-full-node-setup
+    depends_on:
+      build:
+        condition: process_completed_successfully
+    readiness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: echo "true"
+
+  bridge_setup:
+    command: |
+      ./scripts/services/indexer/build
+    availability:
+      restart: exit_on_failure
+    depends_on:
+      setup:
+        condition: process_healthy
