+++ process-compose/bridge/process-compose.yml
@@ -1,0 +1,67 @@
+version: "3"
+
+processes:
+
+  build:
+    command: |
+      exit 0
+
+  setup:
+    command: |
+      # sleep forever
+      sleep 999999999d
+    depends_on:
+      build:
+        condition: process_completed_successfully
+
+  celestia-light-node:
+    command: |
+      exit 1
+    depends_on:
+      setup:
+        condition: process_healthy
+
+  celestia-light-node-synced:
+    command: |
+      wait-for-celestia-light-node
+    depends_on:
+      celestia-light-node:
+        condition: process_healthy
+
+  m1-da-light-node:
+    command: |
+      m1-da-light-node
+    depends_on:
+      celestia-light-node:
+        condition: process_healthy
+      celestia-light-node-synced:
+        condition: process_completed_successfully
+    readiness_probe:
+      initial_delay_seconds: 3
+      exec:
+        command: grpcurl -plaintext 0.0.0.0:30730 list
+    
+  suzuka-full-node:
+    command: |
+      suzuka-full-node
+    env:
+      RUST_LOG: info,aptos-indexer=debug
+    depends_on:
+      m1-da-light-node:
+        condition: process_healthy
+    readiness_probe:
+      initial_delay_seconds: 10
+      exec:
+        command: curl http://0.0.0.0:30731
+
+  suzuka-faucet: 
+    command : |
+      suzuka-faucet-service run-simple
+    depends_on:
+      suzuka-full-node:
+        condition: process_healthy
+    readiness_probe:
+      initial_delay_seconds: 30
+      exec:
+        command: curl http://0.0.0.0:30732
+
