+++ scripts/services/bridge/build
@@ -1,0 +1,38 @@
+#!/usr/bin/env bash
+set -e
+
+echo "Building m1-da-light-node..."
+cargo build $CARGO_PROFILE_FLAGS -p m1-da-light-node --features "sequencer"
+echo "Built m1-da-light-node!"
+
+echo "Building m1-da-light-node-celestia-appd..."
+cargo build $CARGO_PROFILE_FLAGS --bin m1-da-light-node-celestia-appd
+echo "Built m1-da-light-node-celestia-appd!"
+
+echo "Building m1-da-light-node-celestia-bridge..."
+cargo build $CARGO_PROFILE_FLAGS --bin m1-da-light-node-celestia-bridge
+echo "Built m1-da-light-node-celestia-bridge!"
+
+echo "Building suzuka-config..."
+cargo build $CARGO_PROFILE_FLAGS --bin suzuka-full-node-setup
+echo "Built suzuka-config!"
+
+echo "Building suzuka-full-node..."
+cargo build $CARGO_PROFILE_FLAGS -p suzuka-full-node
+echo "Built suzuka-full-node!"
+
+echo "Building suzuka-faucet-service..."
+cargo build $CARGO_PROFILE_FLAGS -p suzuka-faucet-service
+echo "Built suzuka-faucet-service!"
+
+echo "Building suzuka-full-node-setup..."
+cargo build $CARGO_PROFILE_FLAGS -p suzuka-full-node-setup
+echo "Built suzuka-full-node-setup!"
+
+echo "Building wait-for-celestia-light-node..."
+cargo build $CARGO_PROFILE_FLAGS --bin wait-for-celestia-light-node
+echo "Built wait-for-celestia-light-node!"
+
+echo "Building Bridge..."
+cargo build $CARGO_PROFILE_FLAGS -p bridge-service
+echo "Built Bridge!"
